backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
aiofiles>=23.1.0
PyPDF2>=3.0.0
google-genai>=0.5.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
pydantic>=1.10.0
requests>=2.28.0
